fixed tests and optimized test-Setup
