+++ jest.config.ts
@@ -6,25 +6,11 @@ import type { Config } from 'jest';
  */
 
 const customJestConfig: Config = {
-    collectCoverage: true,
-    coverageProvider: 'v8',
-    coverageReporters: ['html-spa','text'],
-    coveragePathIgnorePatterns: [
-        "**/_playwrightConfig/**",
-        "**/_jestConfig/**"
-    ],
     maxWorkers: 1,
     projects: [
-        "<rootDir>/jest.frontend.config.ts",
         "<rootDir>/jest.backend.config.ts",
+        "<rootDir>/jest.frontend.config.ts",
     ],
-    reporters: [
-        "default",
-        ["./node_modules/jest-html-reporter", {
-            "pageTitle": "Test Report"
-        }]
-    ],
-    testResultsProcessor: "./node_modules/jest-html-reporter"
 };
 
 module.exports = customJestConfig;
